requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
python-dotenv==1.0.0
python-engineio==3.14.2
python-socketio==4.6.1
six==1.16.0
websockets==11.0
yarl==1.8.2
